Prefactor: rename localscanner to securedclustercertgen, extract common code
